Compliance updates #1516
