+++ Library_Engine/Query/Names.cs
@@ -38,22 +38,6 @@ namespace BH.Engine.Library
         {
             return Library(libraryName).Select(x => x.Name).ToList();
         }
-
-        /***************************************************/
-
-        public static List<string> LibraryNames()
-        {
-            return LibraryStrings().Keys.ToList();
-        }
-
-        /***************************************************/
-
-        public static List<string> AllLibraryPathNames()
-        {
-            return LibraryPaths().Keys.ToList();
-        }
-
-        /***************************************************/
     }
 }
